yの方をintからdoubleに修正

+++ main.c
@@ -4,7 +4,7 @@
 
 int myRoot(double x) // x の平方根を返す関数。x >= 0 ではないとき exit(1) とする。 
 {
-	int y;
+	double y;
 	if( x < 0 ){
 		exit(1);
 	}
